Added requirements.txt
requirements.txt is generated by doing `pip freeze > requirements.txt`.
It specifies the packages as well as the package version required by
this project.

The packages are specified with exact version requirements (e.g.
Jinja2==3.1.2) so I may change it later on to be more permissive - maybe
something like Jinja2>=3.0. We'll see.