Added Post class.
This is the main structure which will be used to store posts. These are
Markdown text files which contain the content to put on a website.

I've chosen to use markdown.Markdown.convert rather than .convertFile
because I want to be able to render the HTML to a template before
writing it out to a file.