Added meta extension to md_reader
The meta extension is pretty useful - it allows me to specify things
like the title and date of a post in the Markdown file without that data
appearing in the actual text. A post will look something like this:

```
title: [post title]
date: [post date]

[actual content]
```

The only issue is that the converter is simple - it assumes that all
files will contain the title and date metadata (not all posts will have
the date metadata).