Added try-catch statement to handle date metadata
md_reader.Meta is just a dictionary, so if we get a KeyError when trying
to retrieve a value, then we know that that the key-value pair isn't in
the dictionary and therefore that piece of metadata doesn't exist in the
file. It isn't an elegant solution but it works.