Added render_html method to Post class
render_html uses jinja2.Template to render the partial HTML fragment in
Post.content to a full HTML file that can be written to disk. It uses 2
templates:
- blog.html (for regular blog posts)
- meta.html (for meta posts - e.g. an about page)

jinja2.Template.render is pretty neat in that if you give it a key-value
pair that it has no match for in the template, it will automatically
pass over it when rendering/processing. This means I only have to do one
call to template.render regardless of the type of template I am actually
using.