Fixed formatting for top level docstring of module
Turns out there's actually a format to the top level docstring of a
module. It goes something like this:
```
"""
[module title]

[module description]
"""
```

Which corresponds to the following in the modules documentation (i.e.
`help(module)`):
```
Help on module [module name]:

NAME
    [module name] - [module title]

DESCRPITION
    [module description]

[whatever else]
```

It should be noted that [module name] and [module title] are two
different things:
- [module name] is usually the filename of the module without the
  extension, e.g. daisy
- [module title] is a programmer defined...thing that is a very short
  description of what the module is

And then of course you have the module description and the docstrings
for the actual contents of the module...it's all very long. :D