Added add_post_to_index_file function to add posts
Adds a single post to the index file (similar to get_post). Checks by
filename first to make sure that the post isn't already in the index
file - I chose filename because if I might reuse titles, but I won't
reuse filenames as I'd be overwriting a post.

Also file.truncate only starts truncating from where you are. So if
you've read a file with file.read and then do file.truncate, it'll chop
off the end of the file...which is nothing. So you need to seek back to
the beginning, and *then* truncate (and do whatever else).