Added check for template directory to check_dirs
I'm sorting out things like edge cases now - where the user might run
the program in a directory where there's nothing to render. This is the
first one that I thought of after trying to run the program in my home
directory and it crashed because it couldn't find the template
directory. Technically, this does the same, just *earlier*.