Added setup.py
There used to be a time when I thought that packaging and distributing
software with Python was easy. Or at least, not as hard as everyone made
it out to be. "Surely they're exaggerating!", I thought. "Surely it's
not *that* bad!", I thought.

Boy was I wrong.

There are a billion different ways to do things, and all of them are
apparently wrong or outdated. You have setup.py, which is a Python
script to specify how to build/install...whatever it is you're making.
And then people were like "Tony, my boy, you're doing it all wrong! Look
at how Frankie does things" and Frankie's got this thing called
setup.cfg, which is a *static* build script thing, and it's
*declarative* so you don't put all your logic in a dynamic file (what
were you thinking Tony, huh? What's the matter with you?) and that's
better style because...reasons? I don't know. But then Paulie comes
along, he shoots Frankie in the foot with pyproject.toml and goes "So he
got shot in the foot, what is it a big deal?", so now everyone has to
use pyproject.toml because *that's* the right one (even though it kind
of still doesn't work and also relies on the other two...?). I don't
know.

Rambling and Goodfellas aside, from what I understand pyproject.toml is
the way forward, but it's still acceptable to use setup.py and most
projects do. I don't know what I will be doing but this seems to work (I
ran pip install . and now I have daisy in my path which is pretty neat)
so I guess I'll stick with this.

(Also I'm tired and I watched Goodfellas 2 nights ago so that's why
Frankie and Paulie and Tommy showed up).