Added installation instructions to README
It's not a good set of instructions. Ideally someone should just be able
to do `pip install daisy` or something and be done with it - i.e. daisy
should be on PyPI but I'm reluctant to do so for a few reasons. That
being said, these instructions *should* work and to be completely
honest, there are much better frameworks and tools for generating static
sites than this. Seriously, go use Hugo or something instead.