Added read_config_file to read configuration files
I'm going to move most of the configuration to a separate file. I'm
using configparser (which is provided by Python) so the configuration
file will use the INI format. I haven't (yet) changed the rest of the
code to actually *use* the new configuration file (it still relies on
the old hardcoded values at/near the top of the file) but will do soon.