Changed read_config_file to use json
So that lasted all of what...2 seconds? The INI format doesn't really
handle lists well so I decided to just use the json library instead. I
don't know if JSON is supposed to be used for configuring programs
(usually it's just used to pass generic descriptions of objects around)
but it'll do for this use case.