Added extra clause to copy_contents_files
So apparently if the file/directory already exists that's an issue.
Which in a way is actually pretty good because I'm not going to be
changing the contents directory super often so all it means is that it's
not constantly/wastefully copying files over everytime I run it, which
is good! I hope.