Added -e postprocessing argument
I actually changed quite a bit. Here's a brief summary:
- I changed `do_postprocessing` to `run_commands`. It now takes a single
  argument which tells `subprocess` what to actually run.
- I replaced the `-c` option because I changed my mind on having a
  subparser, it seemed too complicated for what I wanted to do.
- I added a new argument: `-e`/`--execute`, which takes commands to
  execute as an argument to itself.
- I changed the help information for `-p`/`--postprocessing` slightly.

Overall, I think I'm happier with this choice. Initially I was going for
something like:
- `daisy ... -p` would run postprocessing from `config.json`.
- `daisy ... -p [commands]` would run postprocessing using `[commands]`.

The downside of that is that if you wanted to do *both*...well you
can't. The current setup is as follows:
- `daisy ... -p` will run postprocessing from `config.json` as usual.
- `daisy ... -e [commands]` will run postprocessing from `[commands]`.

Note that these aren't mutually exclusive, so if you wanted to you could
use both.

I think, in terms of things left to do:
- Add options to modify command line output (`-q`: quiet, `-v`:
  verbose). I'll probably just add one or the other, not both.
- Start breaking up the main file into seperate parts. This is a bit
  iffy given how much of the code depends on the `CONFIG` global
  variable (damnit, they were right!) and to be honest, maybe it's not
  even necessary (the entire program is under 500 SLOC including
  comments and docstrings).
- Testing. Oh god I'm not looking forward to this because I have no clue
  how to test **any of this!** How *are* you supposed to test a CLI
  application? None of the functions return anything!
- Actually release this on PyPI or something so people don't follow my
  crappy instructions of `git clone lol`.
- Maybe change how templating works, so I can give this to someone else
  and they can configure it to do something else rather than just, you
  know, blog posts and the index/about pages. Maybe a seperate projects
  page/area/thing? This is probably the most difficult and something I
  won't do until I need it, so yeah.

Other than that, this program is pretty much complete! I really don't
want to add more than what I have simply because there's no reason to.
Some of the things I have already are a bit unnecessary (I definitely
won't be using `-e` or `-c`) and I've already feature crept a *lot* so I
want to avoid any more of that. Overall, it's been a fun journey, I feel
like I learnt a lot. :)