Added quiet mode to daisy and new function
The new function is `output_info` - takes a single argument (message),
outputs it only if the `quiet` key in the `CONFIG` dictionary is false.
May need to change things slightly given that `CONFIG` also pulls things
from `config.json` so it's possible to add the key there, but it'll be
ignored in favour of the flag being set (or not). Need to add that but
I'm too tired haha.