Changed code relating to setting quiet output
`quiet` is now a required key in `config.json`. I figured I'd add an
`or` statement to the assignment variable to either set the
configuration for quiet output to what it's currently set to or to what
the command line flag is. If the config says quiet and nothing is
specified at runtime, then the output is quiet. If the config says not
quiet and nothing is specified at runtime, the output is loud. If the
options are the same then it's also fine. It'll break if there's nothing
in the config though, regardless of flag or not.