Added flag to copy existing directories in copy_contents_directory
Hello. It's been a while.

I tried messing around with this to build my website (again) and
realised that the `.contents` directory wasn't being copied over,
because:
1. shutil.copytree doesn't (normally) allow copying existing
   directories.
2. I was catching the exception that shutil.copytree throws when copying
   existing directories so I didn't know what was going on.

Which is just classic me. Anyway, this commit adds a flag to copy over
existing files while also adding new files, which is what I wanted.